% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{dist_calc}
\alias{dist_calc}
\title{Calculate 2D and Cumulative Distance Between Coordinates}
\usage{
dist_calc(df, longlat = TRUE)
}
\arguments{
\item{longlat}{logical for calculating Great Circle (TRUE) or Euclidean (FALSE) distances. Defaults to TRUE.
Great Circle distance is returned in km.}

\item{x}{data frame containing coordinate data as columns labeled "Longitude" (x) and "Latitude" (y).}
}
\description{
Appends two columns to a data frame containing Longitude and Latitude coordinates. The first column (Dist) is the
incremental change in distance between coordinates. The second column (DistTotal) is the cumulative distance along a
2D transect beginning at 0.
}
\details{
Relies on the spDistsN1 function from the package 'sp.'
}
\examples{
df_dist <- dist_calc(df, longlat = TRUE)
}
\keyword{circle,}
\keyword{dist}
\keyword{great}
