% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Value Negating}
\usage{
... \%nin\% NA
}
\arguments{
\item{x}{vector or NULL: the values to be matched.}

\item{y}{vector or NULL: the values to be matched against.}
}
\description{
%nin% is the opposite of the binary operator %in%. It returns a logical vector
indicating if there is a mismatch (TRUE) or a match (FALSE). In other words, it
returns TRUE for the values in the first vector but not in the second.
}
\examples{
x = 1:6
y = 4:10
x \%nin\% y # which values of x are not in y?
# TRUE  TRUE  TRUE FALSE FALSE FALSE
}
\keyword{same}
