% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{same}
\alias{same}
\title{Test if All Vecter Elements are Equal}
\usage{
same(x)
}
\arguments{
\item{x}{Vector to be tested.}
}
\description{
same(x) is a simple function to check if all of the elements of a vector are equal. NA's are ignored using na.omit.
}
\examples{
x <- rep(3, times = 10)
same(x) # returns TRUE
y <- 1:10
same (y) # returns FALSE
}
\keyword{same}
