% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{str2lfmm}
\alias{str2lfmm}
\title{Prepare Input Files for Latent Factor Mixed Modeling (LFMM)}
\usage{
str2lfmm(str.data, trait.data, exclude)
}
\arguments{
\item{str.data}{SNP data in STRUCTURE format}

\item{trait.data}{Phenotype or environment data}

\item{exclude}{Vector of sample IDs to exclude}
}
\description{
str2lfmm recodes biallelic SNP loci in STRUCTURE format to a format suitable for LFMM.
An arbitrary vector of reference alleles is automatically chosen, though the reference
can be specified by the user if known. str2lfmm accepts both diploid and haploid SNP data.
Each diploid individual must be coded on two lines. Ambiguity codes must be removed and
nucleotides must be coded as integers (e.g. A, T, C, G as 1, 2, 3, 4). Trait data (either
phenotype or environmental variables) can also be included. Samples may be excluded by
specifying using 'exclude,'
}
\examples{
str2lfmm(str.data = str.data,
trait.data = trait.data,
exclude = NULL)
}
\keyword{lfmm,}
\keyword{recode}
\keyword{structure,}
